refactor: Adapting the class and the main file
